+++ apps/client/src/page/handoverNote/component/Custom.tsx
@@ -2,14 +2,14 @@ import { Button, Label, scrollStyle } from '@tiki/ui';
 
 import { SetStateAction, useState } from 'react';
 
+import File from '@/page/handoverNote/component/File/File';
 import {
   fileBoxStyle,
+  labelStyle,
   layoutStyle,
   noteWrapperStyle,
   textareaStyle,
-} from '@/page/handoverNote/component/Custom/Custom.style';
-import File from '@/page/handoverNote/component/File/File';
-import { labelStyle } from '@/page/handoverNote/component/Template/Template.style';
+} from '@/page/handoverNote/component/style';
 import { CustomNoteData } from '@/page/handoverNote/type/note';
 
 import { FileType } from '@/shared/component/FileImportModal/FileImportModal';
