+++ apps/client/src/page/handoverNote/component/NoteInfo.tsx
@@ -10,7 +10,7 @@ import {
   infoStyle,
   plusBtnStyle,
   titleStyle,
-} from '@/page/handoverNote/component/NoteInfo/NoteInfo.style';
+} from '@/page/handoverNote/component/style';
 import { CreateNoteInfoType } from '@/page/handoverNote/type/note';
 
 import { resizeTextarea } from '@/page/handoverNote/util/resizeTextarea';
