+++ apps/client/src/page/handoverNote/component/style.ts
@@ -1,5 +1,74 @@
+
 import { css } from '@emotion/react';
 import { theme } from '@tiki/ui';
+
+export const noteWrapperStyle = css({
+  display: 'flex',
+
+  maxHeight: 'calc(100vh - 33rem)',
+
+  flexDirection: 'column',
+  gap: '4rem',
+
+  paddingBottom: '5rem',
+
+  overflowY: 'scroll',
+});
+
+
+
+export const textareaStyle = css({
+  width: '100%',
+  height: '40.2rem',
+
+  padding: '2.1rem 2rem 2.1rem 2rem',
+
+  borderRadius: '1rem',
+  border: `1px solid ${theme.colors.gray_300}`,
+
+  resize: 'none',
+
+  '&:focus': {
+    boxShadow: theme.shadow.inset_focus,
+
+    transition: '0.2s ease-in-out',
+
+    outline: 'none',
+  },
+  '::placeholder': {
+    color: theme.colors.gray_500,
+
+    ...theme.text.body07,
+  },
+});
+
+
+
+export const layoutStyle = css({
+  display: 'flex',
+  flexDirection: 'column',
+
+  width: '100%',
+
+  gap: '1.6rem',
+
+  '& > label': {
+    color: theme.colors.gray_800,
+  },
+});
+
+export const labelStyle = css({
+  color: theme.colors.gray_800,
+});
+
+export const fileBoxStyle = css({
+  display: 'flex',
+
+  flexWrap: 'wrap',
+
+  gap: '1.6rem',
+});
+
 
 export const entireInfoStyle = css({
   display: 'flex',
